Work with temp file for video encoding
Encode video to a temporary file and rename the file afterwards. In case
encoding fails or is interrupted, on a next run the temp file will be
removed. This makes it possible to "resume" the process or at least
detect which files have been completely converted and which only partly.