Honor 'use_orig' for videos
When 'use_orig' is True, copy the original video files (or symlink them,
depending if 'orig_link' is True) to the destination, and use them as they are,
without processing.

This setting is ignored for all files not supported in HTML5; those are
converted anyway, because otherwise we could not reproduce them.