use abbr to explain what the metadata fields are
Keep the normal font for browsers that switch to smallcaps, as
suggested by MDN:

https://developer.mozilla.org/en-US/docs/Web/HTML/Element/abbr

We remove the dotted line normalization currently configured in the
`normalize.css` file because it actually creates a *second* dotted
line in Firefox 57. The directive there is `text-decoration` which is
what we disable here instead. It looks better without any dotted line.

We leave the trailing commas and strings like `ISO`, `f/` or `mm` but
those could also be managed through CSS. `<br>` tags may similarly be
managed better through CSS as well.