Add IPTC Headline (2:105) and  to iptc_data
According to iptc.org, the 'title' field (2:05) is "a shorthand
reference for the digital image" and may (User Note) be the filename,
whereas the IPTC Headline is defined as "A brief synopsis of the
caption.".

Thus the headline should be used by sigal to provide a short
description, instead of the title tag.